first pass at moving things around
