+++ lib/realms/effects.rb
@@ -22,10 +22,9 @@ module Realms
         end
 
         effect(:discard) do |amount|
-          sub = turn_structure.on_beginning_of(phase: :main, forx: passive_player) do
+          turn_structure.once("main:beginning", player: passive_player) do
             choose(active_player.hand.cards) do |chosen_card|
               active_player.discard(chosen_card)
-              sub.cancel
             end
           end
         end
