+++ lib/realms/phases.rb
@@ -1,11 +1,3 @@
-require "realms/phases/phase"
-require "realms/phases/setup"
-require "realms/phases/upkeep"
-require "realms/phases/main"
-require "realms/phases/discard"
-require "realms/phases/draw"
-require "realms/refactor"
-
 module Realms
   module Phases
   end
