+++ lib/realms/turns/declarations.rb
@@ -1,0 +1,9 @@
+require "realms/turns/declarations/phase"
+require "realms/turns/declarations/loop"
+
+module Realms
+  module Turns
+    module Declarations
+    end
+  end
+end
